Streamline author_name_string()
Do we really need Optional everywhere?  If I am asking for a Person's name, I ALWAYS expect to get a str back.  That str can be "" if the name is not known but a name is a str.  Similarly, a bool is either True or False -- That is the definition of a bool.  An int is a number...  Is there some database or Pydantic reason for using Optional everywhere?

Without the Optional bits, this function would be even more streamlined...
```
def author_name_string(self) -> str:
        return self.name_string.strip() or f"{self.given} {self.surname}".strip()
```